Fix lint error...

+++ internal/services/mssql/mssql_database_data_source.go
@@ -134,7 +134,7 @@ func dataSourceMsSqlDatabaseRead(d *pluginsdk.ResourceData, meta interface{}) er
 
 			maxSizeGb := int64(0)
 			if props.MaxSizeBytes != nil {
-				maxSizeGb = int64((pointer.From(props.MaxSizeBytes)) / int64(1073741824))
+				maxSizeGb = (pointer.From(props.MaxSizeBytes)) / int64(1073741824)
 			}
 			d.Set("max_size_gb", maxSizeGb)
